+++ src/views/layouts-wallet/TheSwapLayout.vue
@@ -15,7 +15,6 @@ import TheWrapperWallet from '@/core/components/TheWrapperWallet';
 import ModuleSwap from '@/modules/swap/ModuleSwap';
 const ETH_TOKEN = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee';
 const DAI_TOKEN = '0x6b175474e89094c44da98b954eedeac495271d0f';
-const AMT = '0.1';
 
 export default {
   components: { TheWrapperWallet, ModuleSwap },
@@ -30,7 +29,7 @@ export default {
     },
     amount: {
       type: String,
-      default: AMT
+      default: ''
     }
   }
 };